frontend/pyproject.toml
[project]
name = "attnres-frontend"
version = "0.1.0"
description = "FastAPI chat UI for AttnRes language model inference"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Jibran", email = "jibran.ijaz@gmail.com" }]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "tokenizers>=0.19.0",
    "huggingface-hub>=0.23.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "bcrypt<4.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
attnres-serve = "app.main:run"

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
